moved api/types into types/api
