+++ dashi/src/lib/api/fetchLayout.ts
@@ -1,7 +1,7 @@
 import { type ComponentState } from "@/lib/types/state/component";
 import { type ContribPoint } from "@/lib/types/model/extension";
+import type { ApiOptions, ApiResult } from "@/lib/types/api";
 import { makeUrl, callApi, fetchApiResult } from "./helpers";
-import type { ApiOptions, ApiResult } from "./types";
 
 export async function fetchLayout(
   contribPoint: ContribPoint,
